backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
google-generativeai==0.8.2
gtts==2.5.3
python-dotenv==1.0.1